+++ database/migrations/2024_10_18_224638_create_purchase_details_table.php
@@ -17,13 +17,13 @@ class CreatePurchaseDetailsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('purchase_id');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('cost', 10, 2);
             $table->decimal('flete_product', 10, 2);
             $table->decimal('flete_total', 10, 2);
             $table->timestamp('deleted_at')->nullable();
             $table->timestamps();
-            
+
             $table->foreign('product_id', 'purchase_details_product_id_foreign')->references('id')->on('products');
             $table->foreign('purchase_id', 'purchase_details_purchase_id_foreign')->references('id')->on('purchases');
         });
