+++ database/migrations/2024_10_18_224640_create_sale_details_table.php
@@ -17,13 +17,13 @@ class CreateSaleDetailsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('sale_id');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('regular_price', 10, 2);
             $table->decimal('sale_price', 10, 2);
             $table->decimal('discount', 10, 2)->default(0.00);
             $table->timestamps();
             $table->timestamp('deleted_at')->nullable();
-            
+
             $table->foreign('product_id', 'sale_details_product_id_foreign')->references('id')->on('products');
             $table->foreign('sale_id', 'sale_details_sale_id_foreign')->references('id')->on('sales');
         });
